Add styles to WelcomeWindow too
